KafkaPython/python/requirements.txt
kafka-python
lz4
crc32c
